feat: wallet abstraction (NGN/USD/STX), verification page, avatar edit, portfolio toast, role selection
- Frontend: FundWalletModal with NGN/USD/STX tabs + Demo Credit button
- Frontend: SendMoneyForm with currency toggle and email/address recipient
- Frontend: VerificationPage at /verification/apply with full form; ProfilePage links to it
- Frontend: Avatar URL editor on own ProfilePage
- Frontend: PortfolioSection success toast after adding work
- Frontend: Creative/Backer role toggle on signup form
- Backend: Auto-create wallet on signup with status='active', currency='NGN'
- Backend: recordSend resolves counterparty by email via users table
- Backend: parseFloat fix for wallet send amounts
- Backend: walletService.createWallet sets active status (not pending)

+++ backend/src/routes/wallets.js
@@ -73,7 +73,7 @@ router.post('/send', requireAuth, async (req, res, next) => {
   try {
     const { user_id, amount, currency, counterparty_user_id, description } = req.body;
     if (!user_id || !amount || !counterparty_user_id) return res.status(400).json({ error: 'user_id, amount, and counterparty_user_id required' });
-    const send = await walletService.recordSend(user_id, { amount: parseInt(amount), currency: currency || 'NGN', counterpartyUserId: counterparty_user_id, description });
+    const send = await walletService.recordSend(user_id, { amount: parseFloat(amount), currency: currency || 'NGN', counterpartyUserId: counterparty_user_id, description });
     if (!send) return res.status(400).json({ error: 'Insufficient balance, wallet not active, or recipient not found' });
     res.status(201).json({ transaction: send, message: 'Send pending. Confirm via /api/wallets/confirm-send after on-chain confirmation.' });
   } catch (err) { console.error('Send error:', err); res.status(500).json({ error: 'Failed to process send' }); }
